Fold api-documenter's ApiJsonFile into api-extractor, eliminating the need to export the ApiJsonGenerator
